% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{empty_as_na}
\alias{empty_as_na}
\title{empty strings as NA}
\source{
https://stackoverflow.com/questions/24172111/change-the-blank-cells-to-na
}
\usage{
empty_as_na(x)
}
\description{
empty strings as NA
}
