% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml2df.R
\name{xml2df}
\alias{xml2df}
\title{string parse that converst xml to text then to df. not the most dependable.}
\usage{
xml2df(nodeset)
}
\arguments{
\item{nodeset}{an xml2 nodeset}
}
\description{
string parse that converst xml to text then to df. not the most dependable.
}
