% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_dig.R
\name{xml_dig}
\alias{xml_dig}
\title{dig into a nodeset until we find top level data. that is, a node of lenght 0. extract any data with node == 0 (indexed TRUE)}
\usage{
xml_dig(nodeset, dig = TRUE)
}
\arguments{
\item{nodeset}{any nodeset of object type xml2}

\item{dig}{option to keep digging beyond the first terminal node. if FALSE then only data from the first set of terminal nodes is returned}
}
\value{
nested lists of dataframes
}
\description{
dig into a nodeset until we find top level data. that is, a node of lenght 0. extract any data with node == 0 (indexed TRUE)
}
