% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_get_paths.R
\name{xml_get_paths}
\alias{xml_get_paths}
\title{view the paths of xml tree. keeps order as one dives into each node. option to mark any terminal paths (paths that end on the last possible node).}
\usage{
xml_get_paths(doc, mark_terminal = ifelse(only_terminal_parent, ">>", NULL),
  only_terminal_parent = FALSE)
}
\arguments{
\item{doc}{an xml doc}

\item{mark_terminal}{string used to mark terminal nodes. NULL by default.}

\item{only_terminal_parent}{whether to return only the parent node of any terminal branch}
}
\description{
view the paths of xml tree. keeps order as one dives into each node. option to mark any terminal paths (paths that end on the last possible node).
}
