% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_get_trees.R
\name{xml_get_trees}
\alias{xml_get_trees}
\title{get a tree of xml simliar to the unix tree command line tool}
\usage{
xml_get_trees(doc, depth = NULL)
}
\arguments{
\item{doc}{an xml doc or nodeset}

\item{depth}{how deep to go. root is 0, next is 1, etc. NULL by default.}
}
\description{
get a tree of xml simliar to the unix tree command line tool
}
