% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_to_df.R
\name{xml_to_df}
\alias{xml_to_df}
\title{function hacked from https://hopstat.wordpress.com/2014/01/14/faster-xml-conversion-to-data-frames/. fastest, most dependable, but requires xpath knowledge.}
\usage{
xml_to_df(file, xpath, is_xml = FALSE, dig = FALSE)
}
\arguments{
\item{file}{xml or zip file imported using xml2}

\item{xpath}{xml path to file}

\item{is_xml}{if already xml, dont parse}

\item{dig}{whether to recursively search into nodes. better and faster to keep FALSE and explicitly write out xpath.}
}
\description{
function hacked from https://hopstat.wordpress.com/2014/01/14/faster-xml-conversion-to-data-frames/. fastest, most dependable, but requires xpath knowledge.
}
