% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml_get_trees.R
\name{xml_view_trees}
\alias{xml_view_trees}
\title{view xml simliar to the unix tree command line tool}
\usage{
xml_view_trees(tree, ...)
}
\arguments{
\item{tree}{an object of class "xml_tree", "xml_tree_list", or "xml_document"}
}
\description{
view xml simliar to the unix tree command line tool
}
